+++ src/main/java/ru/vyarus/java/generics/resolver/util/walk/TypesWalker.java
@@ -40,7 +40,7 @@ import java.util.Map;
  * <p>
  * NOTE: all variables in type ({@link java.lang.reflect.TypeVariable}) will be replaced in types into
  * lower bound. If you need to preserve variables (e.g. to match variable), make them explicit with
- * {@link GenericsUtils#preserveVariables(Type)}.
+ * {@link ru.vyarus.java.generics.resolver.util.TypeVariableUtils#preserveVariables(Type)}.
  *
  * @author Vyacheslav Rusakov
  * @since 11.05.2018