+++ src/test/groovy/ru/vyarus/java/generics/resolver/cases/order/Base.java
@@ -1,0 +1,8 @@
+package ru.vyarus.java.generics.resolver.cases.order;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 17.12.2018
+ */
+public class Base<K> {
+}