+++ src/test/groovy/ru/vyarus/java/generics/resolver/cases/order/MyClass.java
@@ -1,0 +1,11 @@
+package ru.vyarus.java.generics.resolver.cases.order;
+
+import java.util.List;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 17.12.2018
+ */
+public class MyClass<T extends List<D>, D> extends Base<D>{
+
+}